requirements.txt
customtkinter
matplotlib
python-dotenv
requests
pyperclip
reportlab
aiohttp